Implemented but still requires integration tests

+++ src/envoy/server/api/depends/default_doe.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from fastapi import Request
 
 from envoy.server.model.config.default_doe import DefaultDoeConfiguration
@@ -6,10 +8,10 @@ from envoy.server.model.config.default_doe import DefaultDoeConfiguration
 class DefaultDoeDepends:
     """Dependency class for populating the request state default_doe with an instance of DefaultDoeConfiguration"""
 
-    import_limit_active_watts: int
-    export_limit_active_watts: int
+    import_limit_active_watts: Decimal
+    export_limit_active_watts: Decimal
 
-    def __init__(self, import_limit_active_watts: int, export_limit_active_watts: int):
+    def __init__(self, import_limit_active_watts: Decimal, export_limit_active_watts: Decimal):
         self.import_limit_active_watts = import_limit_active_watts
         self.import_limit_active_watts = export_limit_active_watts
